Handle sending and receiving operations more robustly
We can end up with a race condition, where we send an operation, the
timeout fires, and we send the same operation again. The server has no
way to tell that it isn't the same operation, so it repeats
it. Instead, we should always keep track of the operation that's
in-flight, so we will only ever have a single operation in flight at
any time.